write a single query for all buildings report and changes blade for this respective change, some change on migrations and some change on blade for beautification

+++ database/migrations/2020_03_30_170347_create_salary__based_logs_table.php
@@ -19,7 +19,7 @@ class CreateSalaryBasedLogsTable extends Migration
             $table->integer('salary');
             $table->string('month');
             $table->integer('group_id');
-            $table->integer('building_id');
+            $table->bigInteger('building_id');
             $table->timestamps();
         });
     }
